requirements.txt
streamlit
langchain
openai
faiss-cpu
pypdf2
python-dotenv
requests
tqdm
sentence-transformers